trzeci commit, po branchui

+++ src/league/project/Player.java
@@ -1,13 +1,13 @@
 package league.project;
 public class Player extends People{
     private String name;
-    private int pac;
-    private int sho;
-    private int def;
-    private int pas;
-    private int dri;
-    private int phy;
-    private int power;
+    private double pac;
+    private double sho;
+    private double def;
+    private double pas;
+    private double dri;
+    private double phy;
+    private double power;
     private ITeam team;
 
     public Player(String name){
